requirements.txt
numpy==2.2.1
opencv_python==4.10.0.84
Pillow==11.1.0
PyAutoGUI==0.9.54
PyQt6==6.8.0
PyQt6_sip==13.9.1
pyzbar==0.1.9
qrcode==8.0
